Update ModelsGenesis pre-trained weight loading
